+++ custom_components/thessla_green_modbus/config_flow/__init__.py
@@ -19,51 +19,7 @@ from homeassistant.const import CONF_HOST, CONF_NAME, CONF_PORT
 from homeassistant.core import HomeAssistant
 from voluptuous import Invalid as VOL_INVALID
 
-from .config_flow_confirm import (
-    build_confirmation_placeholders as _build_confirmation_placeholders,
-)
-from .config_flow_device_validation import validate_input_impl as _validate_input_impl
-from .config_flow_entry import build_unique_id as _build_unique_id_impl
-from .config_flow_entry import prepare_entry_payload as _prepare_entry_payload_impl
-from .config_flow_errors import classify_os_error as _classify_os_error_impl
-from .config_flow_errors import (
-    should_log_timeout_traceback as _should_log_timeout_traceback_impl,
-)
-from .config_flow_network import looks_like_hostname as _looks_like_hostname_impl
-from .config_flow_options import denormalize_option as _denormalize_option_impl
-from .config_flow_options import normalize_baud_rate as _normalize_baud_rate_impl
-from .config_flow_options import normalize_parity as _normalize_parity_impl
-from .config_flow_options import normalize_stop_bits as _normalize_stop_bits_impl
-from .config_flow_options import strip_translation_prefix as _strip_translation_prefix_impl
-from .config_flow_options_form import build_options_form_payload as _build_options_form_payload
-from .config_flow_payloads import caps_to_dict as _caps_to_dict_impl
-from .config_flow_payloads import normalize_connection_type as _normalize_connection_type_impl
-from .config_flow_reauth import process_reauth_submission as _process_reauth_submission_impl
-from .config_flow_reauth_confirm import apply_reauth_update as _apply_reauth_update_impl
-from .config_flow_runtime import TIMEOUT_EXCEPTIONS
-from .config_flow_runtime import (
-    call_with_optional_timeout as _call_with_optional_timeout_impl,
-)
-from .config_flow_runtime import (
-    is_request_cancelled_error as _is_request_cancelled_error_impl,
-)
-from .config_flow_runtime import load_scanner_module as _load_scanner_module
-from .config_flow_runtime import run_with_retry as _run_with_retry_impl
-from .config_flow_schema import build_connection_schema as _build_connection_schema_impl
-from .config_flow_schema import build_reconfigure_schema as _build_reconfigure_schema_impl
-from .config_flow_steps import extract_discovered_state as _extract_discovered_state_impl
-from .config_flow_steps import merge_options_payload as _merge_options_payload_impl
-from .config_flow_steps import resolve_reauth_entry as _resolve_reauth_entry_impl
-from .config_flow_steps import resolve_reauth_form_state as _resolve_reauth_form_state_impl
-from .config_flow_steps import validate_options_submission as _validate_options_submission_impl
-from .config_flow_user_submit import process_user_submission as _process_user_submission_impl
-from .config_flow_validation import (
-    process_scan_capabilities_bound as _process_scan_capabilities_bound,
-)
-from .config_flow_validation import validate_rtu_config_bound as _validate_rtu_config_bound
-from .config_flow_validation import validate_slave_id as _validate_slave_id_impl
-from .config_flow_validation import validate_tcp_config_bound as _validate_tcp_config_bound
-from .const import (
+from ..const import (
     CONF_SLAVE_ID,
     CONF_TIMEOUT,
     CONNECTION_TYPE_TCP,
@@ -78,9 +34,53 @@ from .const import (
     DOMAIN,
     MAX_BATCH_REGISTERS,
 )
-from .errors import CannotConnect, InvalidAuth
-from .modbus_exceptions import ModbusIOException
-from .options import MODBUS_BAUD_RATES, MODBUS_PARITY, MODBUS_STOP_BITS
+from ..errors import CannotConnect, InvalidAuth
+from ..modbus_exceptions import ModbusIOException
+from ..options import MODBUS_BAUD_RATES, MODBUS_PARITY, MODBUS_STOP_BITS
+from .confirm import (
+    build_confirmation_placeholders as _build_confirmation_placeholders,
+)
+from .device_validation import validate_input_impl as _validate_input_impl
+from .entry import build_unique_id as _build_unique_id_impl
+from .entry import prepare_entry_payload as _prepare_entry_payload_impl
+from .errors import classify_os_error as _classify_os_error_impl
+from .errors import (
+    should_log_timeout_traceback as _should_log_timeout_traceback_impl,
+)
+from .network import looks_like_hostname as _looks_like_hostname_impl
+from .options import denormalize_option as _denormalize_option_impl
+from .options import normalize_baud_rate as _normalize_baud_rate_impl
+from .options import normalize_parity as _normalize_parity_impl
+from .options import normalize_stop_bits as _normalize_stop_bits_impl
+from .options import strip_translation_prefix as _strip_translation_prefix_impl
+from .options_form import build_options_form_payload as _build_options_form_payload
+from .payloads import caps_to_dict as _caps_to_dict_impl
+from .payloads import normalize_connection_type as _normalize_connection_type_impl
+from .reauth import process_reauth_submission as _process_reauth_submission_impl
+from .reauth_confirm import apply_reauth_update as _apply_reauth_update_impl
+from .runtime import TIMEOUT_EXCEPTIONS
+from .runtime import (
+    call_with_optional_timeout as _call_with_optional_timeout_impl,
+)
+from .runtime import (
+    is_request_cancelled_error as _is_request_cancelled_error_impl,
+)
+from .runtime import load_scanner_module as _load_scanner_module
+from .runtime import run_with_retry as _run_with_retry_impl
+from .schema import build_connection_schema as _build_connection_schema_impl
+from .schema import build_reconfigure_schema as _build_reconfigure_schema_impl
+from .steps import extract_discovered_state as _extract_discovered_state_impl
+from .steps import merge_options_payload as _merge_options_payload_impl
+from .steps import resolve_reauth_entry as _resolve_reauth_entry_impl
+from .steps import resolve_reauth_form_state as _resolve_reauth_form_state_impl
+from .steps import validate_options_submission as _validate_options_submission_impl
+from .user_submit import process_user_submission as _process_user_submission_impl
+from .validation import (
+    process_scan_capabilities_bound as _process_scan_capabilities_bound,
+)
+from .validation import validate_rtu_config_bound as _validate_rtu_config_bound
+from .validation import validate_slave_id as _validate_slave_id_impl
+from .validation import validate_tcp_config_bound as _validate_tcp_config_bound
 
 _LOGGER = logging.getLogger(__name__)
 __all__ = ["VOL_INVALID", "CannotConnect", "ConfigFlow", "InvalidAuth", "validate_input"]
