+++ custom_components/thessla_green_modbus/services/dispatch.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from .modbus_exceptions import ConnectionException, ModbusException
+from ..modbus_exceptions import ConnectionException, ModbusException
 
 
 async def write_register(
